refactor(core): consolidate socket listeners and services
- Moved global socket event handling (execution control, plot history) from React hooks to `core/init/socketListeners.ts`, removing `useSessionControlEvents` and `usePlotHistoryEvents`.
- Inlined `sessionControl.ts` service logic into `code.ts` command module.
- Removed `onOpenSettings` props from `StatusBar` and `AIPanel`, using direct command execution.
